+++ src/dave_data/__init__.py
@@ -1,6 +1,8 @@
 __version__ = "v0.0.1"
 
+from .area_to_polygon import federal_state_to_polygon
 from .area_to_polygon import postalcode_to_polygon
+from .area_to_polygon import town_to_polygon
 from .core import compute
 from .core import get_data
 from .datapool.osm_request import osm_request
@@ -10,6 +12,8 @@ from .datapool.read_data import read_postal
 
 __all__ = [
     "postalcode_to_polygon",
+    "federal_state_to_polygon",
+    "town_to_polygon",
     "compute",
     "get_data",
     "osm_request",